First Commit by Raul Maldonado
Changes:

Major refactor of how weights, networks, and layers are stored

.network is now a storage of the values on each node from the last call
of .getOutput

.net_setup is the number of nodes for each layer including the input and
and output

.weights is now a list of arrays that are used to calculate the values
of each node from the previous layer.

For each node, the previous array layer (for example the inputs) is then
then dot producted with the weights array to get the total weighted
value. These are then put through a sigmoid function to determine the
final output